requirements.txt
# Baarn Politiek MCP Server - Dependencies

# Core
requests>=2.28.0
pyyaml>=6.0
python-dotenv>=1.0.0

# MCP Server
mcp>=1.0.0

# REST API Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0

# PDF Processing
pdfplumber>=0.9.0
PyMuPDF>=1.23.0

# Data processing
numpy>=1.24.0

# Document generatie
python-docx>=0.8.11
python-pptx>=0.6.23
openpyxl>=3.1.2
Pillow>=10.0.0

# Image deduplication (perceptual hashing)
imagehash>=4.3.0

# OCR (tekst uit afbeeldingen)
pytesseract>=0.3.10

# Web scraping
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Embeddings (VERPLICHT voor semantic search)
sentence-transformers>=2.2.0
torch>=2.0.0

# Transcriptie (Whisper voor video/audio)
openai-whisper>=20231117
yt-dlp>=2024.1.0
ffmpeg-python>=0.2.0

# CLI progress output
rich>=13.0.0
